+++ yuno_sdk/example/lib/core/helpers/envs.dart
@@ -1,0 +1,7 @@
+final class Envs {
+  static const apiKey = String.fromEnvironment('apiKey');
+  static const appId = String.fromEnvironment('appId');
+  static const messagingSenderId = String.fromEnvironment('messagingSenderId');
+  static const projectId = String.fromEnvironment('projectId');
+  static const storageBucket = String.fromEnvironment('storageBucket');
+}