+++ yuno_sdk_android/android/src/main/kotlin/com/example/yuno_sdk_android/core/utils/keys/Key.kt
@@ -1,7 +1,0 @@
-package com.example.yuno_sdk_android.core.utils.keys
-
-object Key {
-    const val init = "initialize"
-    const val startPaymentLite = "startPaymentLite"
-    const val continuePayment = "continuePayment"
-}