+++ yuno_sdk_android/android/src/main/kotlin/com/example/yuno_sdk_android/features/start_payment_lite/models/PaymentMethod.kt
@@ -1,6 +1,0 @@
-package com.example.yuno_sdk_android.features.start_payment_lite.models
-
-data class PaymentMethod(
-    val vaultedToken:String?,
-    val paymentMethodType:String,
-    )