+++ yuno_sdk_android/android/src/main/kotlin/com/yuno_flutter/yuno_sdk_android/core/utils/keys/Key.kt
@@ -1,0 +1,7 @@
+package com.yuno_flutter.yuno_sdk_android.core.utils.keys
+
+object Key {
+    const val init = "initialize"
+    const val startPaymentLite = "startPaymentLite"
+    const val continuePayment = "continuePayment"
+}