+++ yuno_sdk_android/android/src/main/kotlin/com/yuno_flutter/yuno_sdk_android/features/start_payment_lite/models/StartPaymentLite.kt
@@ -1,0 +1,34 @@
+package com.yuno_flutter.yuno_sdk_android.features.start_payment_lite.models
+data class StartPaymentLite(
+    val countryCode: String,
+    val showPaymentStatus: Boolean,
+    val checkoutSession: String,
+    val paymentMethodSelected: PaymentMethod
+)
+fun Map<String, Any>.toStartPaymentLite(): Result<StartPaymentLite> {
+
+        return try {
+            val countryCode = this["countryCode"] as? String  ?: return Result.failure(IllegalArgumentException("Missing or invalid countryCode"))
+            val checkoutSession = this["checkoutSession"] as? String ?: return Result.failure(IllegalArgumentException("Missing or invalid checkoutSession"))
+            val showPaymentStatus = this["showPaymentStatus"] as? Boolean ?: return Result.failure(IllegalArgumentException("Missing or invalid showPaymentStatus"))
+            val paymentMethodSelected = this["paymentMethodSelected"] as? Map<*, *> ?: return Result.failure(
+                IllegalArgumentException("Missing or invalid paymentMethodSelected")
+            )
+            val vaultedToken = paymentMethodSelected["vaultedToken"] as? String?
+            val paymentMethodType = paymentMethodSelected["paymentMethodType"] as? String  ?: return Result.failure(IllegalArgumentException("Missing or invalid paymentMethodType"))
+
+            Result.success(
+                StartPaymentLite(
+                  countryCode = countryCode,
+                  showPaymentStatus = showPaymentStatus,
+                  checkoutSession = checkoutSession,
+                  paymentMethodSelected = PaymentMethod(
+                      vaultedToken = vaultedToken,
+                      paymentMethodType = paymentMethodType
+                  )
+                )
+            )
+        } catch (e: Exception){
+            Result.failure(e)
+        }
+}