+++ yuno_sdk_android/CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
-
-* TODO: Describe initial release.
+## 0.0.1 - Android
+- feat: Yuno initialization
+- feat: Yuno start checkout
+- feat: Yuno start payment lite
+- feat: Yuno continue payment
+- feat: Yuno create payment
+- feat: Yuno payment result